Make move_and_slide independent of process delta
Adds a  fixed deltaTime of 1/60 to the CharacterBody3D. This allows to
increase the time that passes in the game by increasing the physics
ticks per second of the Engine.

Normally, when increasing physics ticks per second done by the Engine
the value of delta decreases, resulting in the same distance moved per
second by this function. By keeping the delta used in this function
constant and idenpendent from the actual physics ticks per second a
Client may simulate more than 60 physics_process frames in a second.
But, those frames behave as if there were only 60 frames. Causing
move_and_slide to cover a bigger distance than intended.
This is useful whenever the latency is increasing and the Server notices
that the buffer of inputs from a Client is shrinking. The Server may
tell that Client to produce frames at a higher rate to fill up the
input buffer and avoid having to predict the input of that Client.